pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "researchmind"
version = "0.1.0"
description = "Enterprise R&D Intelligence Platform"
requires-python = ">=3.11"
dependencies = [
    # Data fetching
    "boto3>=1.34.0",
    "httpx>=0.27.0",
    # Embeddings and ML
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.13.0",
    # API serving
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "python-multipart>=0.0.9",
    # LLM
    "anthropic>=0.28.0",
    # LangChain retrieval
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-chroma>=0.2.0",
    # Vector store (user docs)
    "chromadb>=0.5.0",
    # Async task queue
    "celery[eventlet]>=5.4.0",
    "eventlet>=0.37.0",
    # Caching
    "redis>=5.0.0",
    # Citation graph
    "networkx>=3.3.0",
    # PDF parsing
    "pypdf>=4.0.0",
    # Experiment tracking and data versioning
    "mlflow>=2.13.0",
    "dvc[s3]>=3.51.0",
    # Utilities
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "arxiv>=3.0.0",
    "ipykernel>=7.2.0",
    "peft>=0.19.1",
    "torch>=2.11.0",
    "bm25s>=0.3.4",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "locust>=2.43.4",
    "adapters>=1.2.0",
    "pymupdf>=1.27.2.2",
    "ragas>=0.4.3",
    "pyvis>=0.3.2",
    "ollama>=0.6.1",
    "instructor>=1.15.1",
    "langsmith>=0.7.33",
    "langgraph>=1.1.8",
    "openai>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/researchmind","api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv.sources]
torch = { index = "pytorch-cuda" }

[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
